requirements.txt
blinker==1.4
boto3==1.20.27
botocore==1.23.27
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.0
s3transfer==0.5.0
six==1.16.0
stripe==2.64.0
urllib3==1.26.7
Werkzeug==2.0.2
flask_emails
# emails
